pyproject.toml
[tool.poetry]
name = "captchanova"
version = "0.1.0"
description = ""
authors = ["blazaid <alberto.diaz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"

[tool.poetry.scripts]
captchanova = "captchanova.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
